pyproject.toml
[project]
name = "design-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-ml>=1.29.0",
    "azure-identity>=1.25.1",
    "azure-keyvault>=4.2.0",
    "azure-search-documents>=11.6.0b1",
    "azure-storage-blob>=12.27.0",
    "fastapi>=0.120.0",
    "openai>=2.5.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.38.0",
]
